+++ src/service/offer/seatReservation.ts
@@ -677,7 +677,7 @@ export function validateAcceptedOffers(params: {
                             throw new factory.errors.Argument('transactionId', 'Movie Ticket payment not accepted');
                         }
                         const movieTicketPaymentAccepted = <factory.seller.IPaymentAccepted<factory.paymentMethodType.MovieTicket>>
-                            movieTheater.paymentAccepted.find((a) => a.paymentMethodType === factory.paymentMethodType.MovieTicket);
+                            movieTheater.paymentAccepted.find((a) => a.paymentMethodType === movieTicket.typeOf);
                         if (movieTicketPaymentAccepted === undefined) {
                             throw new factory.errors.Argument('transactionId', 'Movie Ticket payment not accepted');
                         }
@@ -808,7 +808,14 @@ export function validateAcceptedOffers(params: {
                                 priceCurrency: factory.chevre.priceCurrency.JPY,
                                 valueAddedTaxIncluded: true,
                                 appliesToVideoFormat: '2D',
-                                appliesToMovieTicketType: mvtkTicketCodeIn.kbnKensyu
+                                appliesToMovieTicket: {
+                                    typeOf: factory.chevre.paymentMethodType.MovieTicket,
+                                    serviceType: mvtkTicketCodeIn.kbnKensyu
+                                },
+                                ...{
+                                    // 互換性維持対応
+                                    appliesToMovieTicketType: mvtkTicketCodeIn.kbnKensyu
+                                }
                             }
                         ];