+++ src/service/transaction/returnOrder/potentialActions/refundMovieTicket.ts
@@ -16,8 +16,11 @@ export async function createRefundMovieTicketActions(params: {
     let refundMovieTicketActions: factory.action.trade.refund.IAttributes<factory.paymentMethodType.MovieTicket>[] = [];
     const refundMovieTicket = params.returnOrderActionParams?.potentialActions?.refundMovieTicket === true;
     if (refundMovieTicket) {
-        const movieTicketPaymentMethods = <factory.order.IPaymentMethod<factory.paymentMethodType.MovieTicket>[]>params.order.paymentMethods
-            .filter((p) => p.typeOf === factory.paymentMethodType.MovieTicket);
+        // tslint:disable-next-line:no-suspicious-comment
+        // TODO 利用可能なムビチケ系統決済方法タイプに対して動的にコーディング
+        const movieTicketPaymentMethods = <factory.order.IPaymentMethod<factory.paymentMethodType.MovieTicket>[]>
+            params.order.paymentMethods.filter((p) => p.typeOf === factory.paymentMethodType.MovieTicket
+                || p.typeOf === factory.paymentMethodType.MGTicket);
 
         refundMovieTicketActions =
             await Promise.all(movieTicketPaymentMethods